refactor(infra): remove approval module cycles
