+++ src/infra/exec-approvals-config.ts
@@ -1,10 +1,24 @@
 // Parses and normalizes the persisted exec approval policy.
 import { randomBytes } from "node:crypto";
 import path from "node:path";
-import type { ExecApprovalsFile, ExecAsk, ExecSecurity } from "./exec-approvals-core.js";
-import { normalizeExecApprovals } from "./exec-approvals.js";
+import {
+  normalizeLowercaseStringOrEmpty,
+  normalizeOptionalString,
+  readStringValue,
+} from "@openclaw/normalization-core/string-coerce";
+import { DEFAULT_AGENT_ID } from "../routing/session-key.js";
+import type {
+  ExecApprovalsAgent,
+  ExecApprovalsDefaults,
+  ExecApprovalsFile,
+  ExecAsk,
+  ExecSecurity,
+} from "./exec-approvals-core.js";
+import type { ExecAllowlistEntry } from "./exec-approvals.types.js";
 import { expandHomePrefix, resolveHomeRelativePath } from "./home-dir.js";
 import { isPlainObject } from "./plain-object.js";
+
+const toStringOrUndefined = readStringValue;
 
 function isExecSecurity(value: unknown): value is ExecSecurity {
   return value === "allowlist" || value === "full" || value === "deny";
@@ -61,7 +75,7 @@ export function resolveExecApprovalsTranscriptPath(): string {
 }
 
 export function createFailClosedExecApprovalsFallback(): ExecApprovalsFile {
-  return normalizeExecApprovals({
+  return normalizeExecApprovalsInternal({
     version: 1,
     defaults: {
       security: "deny",
@@ -142,13 +156,180 @@ export function parsePersistedExecApprovals(raw: string): ExecApprovalsFile {
   try {
     const parsed = JSON.parse(raw) as unknown;
     if (isValidPersistedExecApprovals(parsed)) {
-      return normalizeExecApprovals(parsed);
+      return normalizeExecApprovalsInternal(parsed);
     }
   } catch {
     // A partial Windows fallback write is existing state, not a missing policy.
   }
   // Never let malformed persisted state inherit permissive product defaults.
   return createFailClosedExecApprovalsFallback();
+}
+
+function normalizeAllowlistPattern(value: string | undefined): string | null {
+  const trimmed = normalizeOptionalString(value) ?? "";
+  return trimmed ? normalizeLowercaseStringOrEmpty(trimmed) : null;
+}
+
+function mergeLegacyAgent(
+  current: ExecApprovalsAgent,
+  legacy: ExecApprovalsAgent,
+): ExecApprovalsAgent {
+  const allowlist: ExecAllowlistEntry[] = [];
+  const seen = new Set<string>();
+  const pushEntry = (entry: ExecAllowlistEntry) => {
+    const patternKey = normalizeAllowlistPattern(entry.pattern);
+    if (!patternKey) {
+      return;
+    }
+    const key = `${patternKey}\x00${entry.argPattern?.trim() ?? ""}`;
+    if (seen.has(key)) {
+      return;
+    }
+    seen.add(key);
+    allowlist.push(entry);
+  };
+  for (const entry of current.allowlist ?? []) {
+    pushEntry(entry);
+  }
+  for (const entry of legacy.allowlist ?? []) {
+    pushEntry(entry);
+  }
+
+  return {
+    security: current.security ?? legacy.security,
+    ask: current.ask ?? legacy.ask,
+    askFallback: current.askFallback ?? legacy.askFallback,
+    autoAllowSkills: current.autoAllowSkills ?? legacy.autoAllowSkills,
+    allowlist: allowlist.length > 0 ? allowlist : undefined,
+  };
+}
+
+function coerceAllowlistEntries(allowlist: unknown): ExecAllowlistEntry[] | undefined {
+  if (!Array.isArray(allowlist) || allowlist.length === 0) {
+    return Array.isArray(allowlist) ? (allowlist as ExecAllowlistEntry[]) : undefined;
+  }
+  let changed = false;
+  const result: ExecAllowlistEntry[] = [];
+  for (const item of allowlist) {
+    if (typeof item === "string") {
+      const trimmed = item.trim();
+      if (trimmed) {
+        result.push({ pattern: trimmed });
+        changed = true;
+      } else {
+        changed = true; // dropped empty string
+      }
+    } else if (item && typeof item === "object" && !Array.isArray(item)) {
+      const pattern = (item as { pattern?: unknown }).pattern;
+      if (typeof pattern === "string" && pattern.trim().length > 0) {
+        result.push(item as ExecAllowlistEntry);
+      } else {
+        changed = true; // dropped invalid entry
+      }
+    } else {
+      changed = true; // dropped invalid entry
+    }
+  }
+  return changed ? (result.length > 0 ? result : undefined) : (allowlist as ExecAllowlistEntry[]);
+}
+
+function ensureAllowlistIds(
+  allowlist: ExecAllowlistEntry[] | undefined,
+): ExecAllowlistEntry[] | undefined {
+  if (!Array.isArray(allowlist) || allowlist.length === 0) {
+    return allowlist;
+  }
+  let changed = false;
+  const next = allowlist.map((entry) => {
+    if (entry.id) {
+      return entry;
+    }
+    changed = true;
+    return { ...entry, id: crypto.randomUUID() };
+  });
+  return changed ? next : allowlist;
+}
+
+function stripAllowlistCommandText(
+  allowlist: ExecAllowlistEntry[] | undefined,
+): ExecAllowlistEntry[] | undefined {
+  if (!Array.isArray(allowlist) || allowlist.length === 0) {
+    return allowlist;
+  }
+  let changed = false;
+  const next = allowlist.map((entry) => {
+    if (typeof entry.commandText !== "string") {
+      return entry;
+    }
+    changed = true;
+    const { commandText: _commandText, ...rest } = entry;
+    return rest;
+  });
+  return changed ? next : allowlist;
+}
+
+function sanitizeExecApprovalPolicy(
+  policy: ExecApprovalsDefaults | ExecApprovalsAgent | undefined,
+): ExecApprovalsDefaults {
+  const security = toStringOrUndefined(policy?.security)?.trim();
+  const ask = toStringOrUndefined(policy?.ask)?.trim();
+  const askFallback = toStringOrUndefined(policy?.askFallback)?.trim();
+  return {
+    security:
+      security === "deny" || security === "allowlist" || security === "full" ? security : undefined,
+    ask: ask === "off" || ask === "on-miss" || ask === "always" ? ask : undefined,
+    askFallback:
+      askFallback === "deny" || askFallback === "allowlist" || askFallback === "full"
+        ? askFallback
+        : undefined,
+    autoAllowSkills: policy?.autoAllowSkills,
+  };
+}
+
+export function normalizeExecApprovalsInternal(file: ExecApprovalsFile): ExecApprovalsFile {
+  const { path: rawSocketPath, token: rawSocketToken } = file.socket ?? {};
+  const socketPath = rawSocketPath?.trim();
+  const token = rawSocketToken?.trim();
+  const agents = { ...file.agents };
+  const legacyDefault = agents.default;
+  if (legacyDefault) {
+    const main = agents[DEFAULT_AGENT_ID];
+    agents[DEFAULT_AGENT_ID] = main ? mergeLegacyAgent(main, legacyDefault) : legacyDefault;
+    delete agents.default;
+  }
+  for (const [key, agent] of Object.entries(agents)) {
+    const coerced = coerceAllowlistEntries(agent.allowlist);
+    const withIds = ensureAllowlistIds(coerced);
+    const allowlist = stripAllowlistCommandText(withIds);
+    const sanitizedPolicy = sanitizeExecApprovalPolicy(agent);
+    const agentChanged =
+      allowlist !== agent.allowlist ||
+      sanitizedPolicy.security !== agent.security ||
+      sanitizedPolicy.ask !== agent.ask ||
+      sanitizedPolicy.askFallback !== agent.askFallback;
+    if (agentChanged) {
+      agents[key] = {
+        ...agent,
+        allowlist,
+        security: sanitizedPolicy.security,
+        ask: sanitizedPolicy.ask,
+        askFallback: sanitizedPolicy.askFallback,
+      };
+    }
+  }
+  const sanitizedDefaults = sanitizeExecApprovalPolicy(file.defaults);
+  const normalized: ExecApprovalsFile = {
+    version: 1,
+    socket: {
+      path: socketPath && socketPath.length > 0 ? socketPath : undefined,
+      token: token && token.length > 0 ? token : undefined,
+    },
+    defaults: {
+      ...sanitizedDefaults,
+    },
+    agents,
+  };
+  return normalized;
 }
 
 export function mergeExecApprovalsSocketDefaults(params: {
