+++ src/infra/exec-approvals.ts
@@ -1,27 +1,17 @@
 // Manages exec approval policy, allowlist entries, and host targeting.
 import {
-  normalizeLowercaseStringOrEmpty,
-  normalizeOptionalString,
-  readStringValue,
-} from "@openclaw/normalization-core/string-coerce";
-import { DEFAULT_AGENT_ID } from "../routing/session-key.js";
-import {
-  DEFAULT_ASK,
-  DEFAULT_AUTO_ALLOW_SKILLS,
-  DEFAULT_EXEC_APPROVAL_ASK_FALLBACK,
-  DEFAULT_SECURITY,
+  normalizeExecApprovalsInternal,
   resolveExecApprovalsPath,
   resolveExecApprovalsSocketPath,
 } from "./exec-approvals-config.js";
 import type {
-  ExecApprovalsAgent,
-  ExecApprovalsDefaults,
   ExecApprovalsFile,
   ExecApprovalsResolved,
   ExecAsk,
   ExecSecurity,
 } from "./exec-approvals-core.js";
 import { withExecApprovalsReadLock, withExecApprovalsReadLockSync } from "./exec-approvals-lock.js";
+import { resolveExecApprovalsFromFilePrepared } from "./exec-approvals-resolver.js";
 import {
   ensureExecApprovals,
   ensureExecApprovalsSnapshot,
@@ -56,299 +46,13 @@ export {
   withAgentExecApprovalsRemoved,
 } from "./exec-approvals-store.js";
 
-const toStringOrUndefined = readStringValue;
-
-function normalizeAllowlistPattern(value: string | undefined): string | null {
-  const trimmed = normalizeOptionalString(value) ?? "";
-  return trimmed ? normalizeLowercaseStringOrEmpty(trimmed) : null;
-}
-
-function mergeLegacyAgent(
-  current: ExecApprovalsAgent,
-  legacy: ExecApprovalsAgent,
-): ExecApprovalsAgent {
-  const allowlist: ExecAllowlistEntry[] = [];
-  const seen = new Set<string>();
-  const pushEntry = (entry: ExecAllowlistEntry) => {
-    const patternKey = normalizeAllowlistPattern(entry.pattern);
-    if (!patternKey) {
-      return;
-    }
-    const key = `${patternKey}\x00${entry.argPattern?.trim() ?? ""}`;
-    if (seen.has(key)) {
-      return;
-    }
-    seen.add(key);
-    allowlist.push(entry);
-  };
-  for (const entry of current.allowlist ?? []) {
-    pushEntry(entry);
-  }
-  for (const entry of legacy.allowlist ?? []) {
-    pushEntry(entry);
-  }
-
-  return {
-    security: current.security ?? legacy.security,
-    ask: current.ask ?? legacy.ask,
-    askFallback: current.askFallback ?? legacy.askFallback,
-    autoAllowSkills: current.autoAllowSkills ?? legacy.autoAllowSkills,
-    allowlist: allowlist.length > 0 ? allowlist : undefined,
-  };
-}
-
-function coerceAllowlistEntries(allowlist: unknown): ExecAllowlistEntry[] | undefined {
-  if (!Array.isArray(allowlist) || allowlist.length === 0) {
-    return Array.isArray(allowlist) ? (allowlist as ExecAllowlistEntry[]) : undefined;
-  }
-  let changed = false;
-  const result: ExecAllowlistEntry[] = [];
-  for (const item of allowlist) {
-    if (typeof item === "string") {
-      const trimmed = item.trim();
-      if (trimmed) {
-        result.push({ pattern: trimmed });
-        changed = true;
-      } else {
-        changed = true; // dropped empty string
-      }
-    } else if (item && typeof item === "object" && !Array.isArray(item)) {
-      const pattern = (item as { pattern?: unknown }).pattern;
-      if (typeof pattern === "string" && pattern.trim().length > 0) {
-        result.push(item as ExecAllowlistEntry);
-      } else {
-        changed = true; // dropped invalid entry
-      }
-    } else {
-      changed = true; // dropped invalid entry
-    }
-  }
-  return changed ? (result.length > 0 ? result : undefined) : (allowlist as ExecAllowlistEntry[]);
-}
-
-function ensureAllowlistIds(
-  allowlist: ExecAllowlistEntry[] | undefined,
-): ExecAllowlistEntry[] | undefined {
-  if (!Array.isArray(allowlist) || allowlist.length === 0) {
-    return allowlist;
-  }
-  let changed = false;
-  const next = allowlist.map((entry) => {
-    if (entry.id) {
-      return entry;
-    }
-    changed = true;
-    return { ...entry, id: crypto.randomUUID() };
-  });
-  return changed ? next : allowlist;
-}
-
-function stripAllowlistCommandText(
-  allowlist: ExecAllowlistEntry[] | undefined,
-): ExecAllowlistEntry[] | undefined {
-  if (!Array.isArray(allowlist) || allowlist.length === 0) {
-    return allowlist;
-  }
-  let changed = false;
-  const next = allowlist.map((entry) => {
-    if (typeof entry.commandText !== "string") {
-      return entry;
-    }
-    changed = true;
-    const { commandText: _commandText, ...rest } = entry;
-    return rest;
-  });
-  return changed ? next : allowlist;
-}
-
-function sanitizeExecApprovalPolicy(
-  policy: ExecApprovalsDefaults | ExecApprovalsAgent | undefined,
-): ExecApprovalsDefaults {
-  const security = toStringOrUndefined(policy?.security)?.trim();
-  const ask = toStringOrUndefined(policy?.ask)?.trim();
-  const askFallback = toStringOrUndefined(policy?.askFallback)?.trim();
-  return {
-    security:
-      security === "deny" || security === "allowlist" || security === "full" ? security : undefined,
-    ask: ask === "off" || ask === "on-miss" || ask === "always" ? ask : undefined,
-    askFallback:
-      askFallback === "deny" || askFallback === "allowlist" || askFallback === "full"
-        ? askFallback
-        : undefined,
-    autoAllowSkills: policy?.autoAllowSkills,
-  };
-}
-
 export function normalizeExecApprovals(file: ExecApprovalsFile): ExecApprovalsFile {
   const socketPath = file.socket?.path?.trim();
   const token = file.socket?.token?.trim();
-  const agents = { ...file.agents };
-  const legacyDefault = agents.default;
-  if (legacyDefault) {
-    const main = agents[DEFAULT_AGENT_ID];
-    agents[DEFAULT_AGENT_ID] = main ? mergeLegacyAgent(main, legacyDefault) : legacyDefault;
-    delete agents.default;
-  }
-  for (const [key, agent] of Object.entries(agents)) {
-    const coerced = coerceAllowlistEntries(agent.allowlist);
-    const withIds = ensureAllowlistIds(coerced);
-    const allowlist = stripAllowlistCommandText(withIds);
-    const sanitizedPolicy = sanitizeExecApprovalPolicy(agent);
-    const agentChanged =
-      allowlist !== agent.allowlist ||
-      sanitizedPolicy.security !== agent.security ||
-      sanitizedPolicy.ask !== agent.ask ||
-      sanitizedPolicy.askFallback !== agent.askFallback;
-    if (agentChanged) {
-      agents[key] = {
-        ...agent,
-        allowlist,
-        security: sanitizedPolicy.security,
-        ask: sanitizedPolicy.ask,
-        askFallback: sanitizedPolicy.askFallback,
-      };
-    }
-  }
-  const sanitizedDefaults = sanitizeExecApprovalPolicy(file.defaults);
-  const normalized: ExecApprovalsFile = {
-    version: 1,
-    socket: {
-      path: socketPath && socketPath.length > 0 ? socketPath : undefined,
-      token: token && token.length > 0 ? token : undefined,
-    },
-    defaults: {
-      ...sanitizedDefaults,
-    },
-    agents,
-  };
-  return normalized;
-}
-
-function isExecSecurity(value: unknown): value is ExecSecurity {
-  return value === "allowlist" || value === "full" || value === "deny";
-}
-
-function isExecAsk(value: unknown): value is ExecAsk {
-  return value === "always" || value === "off" || value === "on-miss";
-}
-
-function normalizeSecurity(value: unknown, fallback: ExecSecurity): ExecSecurity {
-  return isExecSecurity(value) ? value : fallback;
-}
-
-function normalizeAsk(value: unknown, fallback: ExecAsk): ExecAsk {
-  return isExecAsk(value) ? value : fallback;
-}
-
-type ResolvedExecPolicyField<TValue extends ExecSecurity | ExecAsk> = {
-  value: TValue;
-  source: string | null;
-};
-
-function resolveDefaultSecurityField(params: {
-  field: "security" | "askFallback";
-  defaults: ExecApprovalsDefaults;
-  fallback: ExecSecurity;
-}): ResolvedExecPolicyField<ExecSecurity> {
-  const defaultValue = params.defaults[params.field];
-  if (isExecSecurity(defaultValue)) {
-    return {
-      value: defaultValue,
-      source: `defaults.${params.field}`,
-    };
-  }
-  return {
-    value: params.fallback,
-    source: null,
-  };
-}
-
-function resolveDefaultAskField(params: {
-  defaults: ExecApprovalsDefaults;
-  fallback: ExecAsk;
-}): ResolvedExecPolicyField<ExecAsk> {
-  if (isExecAsk(params.defaults.ask)) {
-    return {
-      value: params.defaults.ask,
-      source: "defaults.ask",
-    };
-  }
-  return {
-    value: params.fallback,
-    source: null,
-  };
-}
-
-function resolveAgentSecurityField(params: {
-  field: "security" | "askFallback";
-  defaults: ExecApprovalsDefaults;
-  agent: ExecApprovalsAgent;
-  rawAgent: ExecApprovalsAgent;
-  wildcard: ExecApprovalsAgent;
-  rawWildcard: ExecApprovalsAgent;
-  agentKey: string;
-  fallback: ExecSecurity;
-}): ResolvedExecPolicyField<ExecSecurity> {
-  const fallbackField = resolveDefaultSecurityField({
-    field: params.field,
-    defaults: params.defaults,
-    fallback: params.fallback,
+  return normalizeExecApprovalsInternal({
+    ...file,
+    socket: { path: socketPath, token },
   });
-  const rawAgentValue = params.rawAgent[params.field];
-  if (rawAgentValue != null) {
-    if (isExecSecurity(params.agent[params.field])) {
-      return {
-        value: params.agent[params.field] as ExecSecurity,
-        source: `agents.${params.agentKey}.${params.field}`,
-      };
-    }
-    return fallbackField;
-  }
-  const rawWildcardValue = params.rawWildcard[params.field];
-  if (rawWildcardValue != null) {
-    if (isExecSecurity(params.wildcard[params.field])) {
-      return {
-        value: params.wildcard[params.field] as ExecSecurity,
-        source: `agents.*.${params.field}`,
-      };
-    }
-    return fallbackField;
-  }
-  return fallbackField;
-}
-
-function resolveAgentAskField(params: {
-  defaults: ExecApprovalsDefaults;
-  agent: ExecApprovalsAgent;
-  rawAgent: ExecApprovalsAgent;
-  wildcard: ExecApprovalsAgent;
-  rawWildcard: ExecApprovalsAgent;
-  agentKey: string;
-  fallback: ExecAsk;
-}): ResolvedExecPolicyField<ExecAsk> {
-  const fallbackField = resolveDefaultAskField({
-    defaults: params.defaults,
-    fallback: params.fallback,
-  });
-  if (params.rawAgent.ask != null) {
-    if (isExecAsk(params.agent.ask)) {
-      return {
-        value: params.agent.ask,
-        source: `agents.${params.agentKey}.ask`,
-      };
-    }
-    return fallbackField;
-  }
-  if (params.rawWildcard.ask != null) {
-    if (isExecAsk(params.wildcard.ask)) {
-      return {
-        value: params.wildcard.ask,
-        source: "agents.*.ask",
-      };
-    }
-    return fallbackField;
-  }
-  return fallbackField;
 }
 
 export type ExecApprovalsDefaultOverrides = {
@@ -460,81 +164,12 @@ export function resolveExecApprovalsFromFile(params: {
 }): ExecApprovalsResolved {
   const rawFile = params.file;
   const file = normalizeExecApprovals(params.file);
-  const defaults = file.defaults ?? {};
-  const agentKey = params.agentId ?? DEFAULT_AGENT_ID;
-  const agent = file.agents?.[agentKey] ?? {};
-  const wildcard = file.agents?.["*"] ?? {};
-  const rawAgent = rawFile.agents?.[agentKey] ?? {};
-  const rawWildcard = rawFile.agents?.["*"] ?? {};
-  const fallbackSecurity = params.overrides?.security ?? DEFAULT_SECURITY;
-  const fallbackAsk = params.overrides?.ask ?? DEFAULT_ASK;
-  const fallbackAskFallback = params.overrides?.askFallback ?? DEFAULT_EXEC_APPROVAL_ASK_FALLBACK;
-  const fallbackAutoAllowSkills = params.overrides?.autoAllowSkills ?? DEFAULT_AUTO_ALLOW_SKILLS;
-  const resolvedDefaults: Required<ExecApprovalsDefaults> = {
-    security: normalizeSecurity(defaults.security, fallbackSecurity),
-    ask: normalizeAsk(defaults.ask, fallbackAsk),
-    askFallback: normalizeSecurity(
-      defaults.askFallback ?? fallbackAskFallback,
-      fallbackAskFallback,
-    ),
-    autoAllowSkills: defaults.autoAllowSkills ?? fallbackAutoAllowSkills,
-  };
-  const resolvedAgentSecurity = resolveAgentSecurityField({
-    field: "security",
-    defaults,
-    agent,
-    rawAgent,
-    wildcard,
-    rawWildcard,
-    agentKey,
-    fallback: resolvedDefaults.security,
-  });
-  const resolvedAgentAsk = resolveAgentAskField({
-    defaults,
-    agent,
-    rawAgent,
-    wildcard,
-    rawWildcard,
-    agentKey,
-    fallback: resolvedDefaults.ask,
-  });
-  const resolvedAgentAskFallback = resolveAgentSecurityField({
-    field: "askFallback",
-    defaults,
-    agent,
-    rawAgent,
-    wildcard,
-    rawWildcard,
-    agentKey,
-    fallback: resolvedDefaults.askFallback,
-  });
-  const resolvedAgent: Required<ExecApprovalsDefaults> = {
-    security: resolvedAgentSecurity.value,
-    ask: resolvedAgentAsk.value,
-    askFallback: resolvedAgentAskFallback.value,
-    autoAllowSkills:
-      agent.autoAllowSkills ?? wildcard.autoAllowSkills ?? resolvedDefaults.autoAllowSkills,
-  };
-  const allowlist = [
-    ...(Array.isArray(wildcard.allowlist) ? wildcard.allowlist : []),
-    ...(Array.isArray(agent.allowlist) ? agent.allowlist : []),
-  ];
-  return {
-    path: params.path ?? resolveExecApprovalsPath(),
-    socketPath: expandHomePrefix(
-      params.socketPath ?? file.socket?.path ?? resolveExecApprovalsSocketPath(),
-    ),
-    token: params.token ?? file.socket?.token ?? "",
-    defaults: resolvedDefaults,
-    agent: resolvedAgent,
-    agentSources: {
-      security: resolvedAgentSecurity.source,
-      ask: resolvedAgentAsk.source,
-      askFallback: resolvedAgentAskFallback.source,
-    },
-    allowlist,
+  return resolveExecApprovalsFromFilePrepared({
+    ...params,
+    rawFile,
     file,
-  };
+    token: params.token ?? file.socket?.token ?? "",
+  });
 }
 
 export {
